+++ tests/test_bruker.py
@@ -9,6 +9,7 @@ from pathlib import Path
 import json
 
 import numpy as np
+import pytest
 
 from .io_for_tests import read_nifti_mrs
 
@@ -17,6 +18,7 @@ bruker_path = Path(__file__).parent / 'spec2nii_test_data' / 'bruker'
 data_path = bruker_path / '20201208_105201_lego_rod_1_3'
 
 
+@pytest.mark.skip(reason="Awaiting fix of brukerapi package")
 def test_fid(tmp_path):
 
     subprocess.check_call(['spec2nii', 'bruker',
@@ -94,6 +96,7 @@ def test_fid(tmp_path):
     assert 'acqp' in hdr_ext
 
 
+@pytest.mark.skip(reason="Awaiting fix of brukerapi package")
 def test_2dseq(tmp_path):
 
     subprocess.check_call(['spec2nii', 'bruker',